Narrow second mt_evtLoadSave f0 return local

+++ src/Event/mt_evtLoadSave.c
@@ -1366,7 +1366,7 @@ void FUN_0038f0f0(u64 param_1)
 
   u32 uVar10;
 
-  u64 uVar11;
+  u32 uVar11;
 
   u64 uVar12;
